chore(xrpl-types): move HexTxHash (de)serializer in separate file
