+++ packages/xrpl-types/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod error;
 pub mod hex_option;
 pub mod msg;
+pub mod tx_hash_hex;
 pub mod types;
